requirements.txt
anthropic>=0.39.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
playwright>=1.40.0
fpdf2>=2.7.0
flask>=3.0.0
